src/content_service/requirements.txt
sqlalchemy==2.0.28
alembic==1.13.1
databases==0.9.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
grpcio==1.58.0
grpcio-tools==1.58.0
fastapi==0.104.1
